added readme and libs for metrics

+++ src/main/kotlin/com/example/dnsService/model/DomainRecord.kt
@@ -1,7 +1,9 @@
 package com.example.dnsService.model
 
+import java.math.BigInteger
+
 data class DomainRecord(
     val domain: String,
     val values: List<String>,
-    val ttl: Int
+    val ttl: BigInteger
 )
